Fix RequestContext
