+++ src/Builder/Pdf/UrlPdfBuilder.php
@@ -6,12 +6,10 @@ use Sensiolabs\GotenbergBundle\Builder\AbstractBuilder;
 use Sensiolabs\GotenbergBundle\Builder\Attributes\NormalizeGotenbergPayload;
 use Sensiolabs\GotenbergBundle\Builder\Attributes\WithBuilderConfiguration;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\ChromiumPdfTrait;
+use Sensiolabs\GotenbergBundle\Builder\Behaviors\Dependencies\RequestContextAwareTrait;
 use Sensiolabs\GotenbergBundle\Builder\BuilderAssetInterface;
 use Sensiolabs\GotenbergBundle\Builder\Util\NormalizerFactory;
 use Sensiolabs\GotenbergBundle\Exception\MissingRequiredFieldException;
-use Symfony\Component\Routing\RequestContext;
-use Symfony\Contracts\Service\Attribute\SubscribedService;
-use Symfony\Contracts\Service\ServiceSubscriberTrait;
 
 /**
  * @see https://gotenberg.dev/docs/routes#url-into-pdf-route
@@ -20,7 +18,7 @@ use Symfony\Contracts\Service\ServiceSubscriberTrait;
 final class UrlPdfBuilder extends AbstractBuilder implements BuilderAssetInterface
 {
     use ChromiumPdfTrait;
-    use ServiceSubscriberTrait;
+    use RequestContextAwareTrait;
 
     public const ENDPOINT = '/forms/chromium/convert/url';
 
@@ -47,19 +45,6 @@ final class UrlPdfBuilder extends AbstractBuilder implements BuilderAssetInterfa
         $this->getBodyBag()->set('route', [$name, $parameters]);
 
         return $this;
-    }
-
-    #[SubscribedService('.sensiolabs_gotenberg.request_context', nullable: true)]
-    protected function getRequestContext(): RequestContext|null
-    {
-        if (
-            !$this->container->has('.sensiolabs_gotenberg.request_context')
-            || !($requestContext = $this->container->get('.sensiolabs_gotenberg.request_context')) instanceof RequestContext
-        ) {
-            return null;
-        }
-
-        return $requestContext;
     }
 
     protected function getEndpoint(): string
